// src/motion-provider/types.ts
export interface MotionMovieAnimationsProps
  extends Omit<MotionAnimationProps, "mode"> {
  /**
   * @description
   * Enter animations are covering the start point
   * of the animation process per slide which
   * means the user will see each slide within
   * the enter animations.
   *
   * @default undefined
   * @type {AnimationKeys[] | AnimationKeys}
   */
  enter: AnimationKeys[] | AnimationKeys;
  /**
   * @description
   * Exit animations are covering the end point
   * of the animation process per slide which
   * means the user will end seeing the slide with
   * the exit prop's animations.
   *
   * @default undefined
   * @type {AnimationKeys[] | AnimationKeys}
   */
  exit: AnimationKeys[] | AnimationKeys;
}

// Core

type GeneralHTMLAttributes = Omit<HTMLAttributes<HTMLElement>, "children">;

export interface MotionContainerProps extends GeneralHTMLAttributes {
  /**
   * @description
   * Defines properties that can be
   * mandatoryly used across MP components. It includes animation modes,
   * transitions, delays, and durations. Basically everything you need to
   * make the web better :)
   *
   * @property {AnimationKeys | AnimationKeys[]} mode - animation mode(s)
   * @property {TransitionKeys} transition - animation transition type
   * @property {number | undefined} [delay] - animation delay
   * @property {number} [duration] - animation duration
   *
   * @example
   *
   * <MotionContainer
   *   elementType="div"
   *   animation={{
   *     mode: ["fadeIn", "filterBlurIn"],
   *     transition: "smooth",
   *   }}
   *   className="your-css-goes-here"
   * />
   *
   */
  animation: MotionAnimationProps;
  /**
   * @description
   * The elementType prop allows you to specify the HTML element type
   * that will be used as the root element for the MotionContainer.
   *
   * @default "div"
   * @type {React.ElementType}
   */
  elementType: React.ElementType;
  /**
   * @description
   * The children prop is a React node that will be rendered inside
   * the MotionContainer component. It is optional and can be used to
   * add content to the component.
   *
   * @default undefined
   * @type {React.ReactNode}
   */
  children?: React.ReactNode;
  /**
   * @description
   * This is the central scaffold part that you might see as
   * a god of CAS(Centralized Animation System). It's actually
   * standing on the top of each MP components to control
   * and manage the animation process by managing the flow.
   *
   * Highly recommended to use with both
   * @type {UseAnimationControlProps}
   * and @type {UseAnimationProps}
   *
   * @default undefined
   * @typedef {Object} MotionControllerProps
   * @param {boolean} isAnimationStopped
   * @param {boolean} reverseAnimation
   * @param {Partial<UseInViewOptions>} [configView]
   * @param {boolean} [trigger]
   *
   * @example
   *
   * //full example
   *
   * const { onReverse, control } = useAnimationControl();
   * const { isAnimationStopped, reverse } = useAnimation(control);
   *
   * // trigger the animation freely
   * <button onClick={onReverse}>Reverse</button>
   * <MotionContainer
   *   elementType="div"
   *   animation={{
   *     mode: ["fadeUp", "translate3dIn"],
   *     transition: "cubicBounce",
   *     duration: 1,
   *   }}
   *   controller={{
   *     trigger: true, // immediately start animation bypassing configView
   *     configView: { once: true, amount: 0.5 } // determine whether to start animation when it's in viewport
   *     isAnimationStopped, // pass the CAS props
   *     reverse // pass the CAS props
   *   }}
   *   className="your-css-goes-here"
   * />
   *   <IfThereIsChildComponent />
   * </MotionContainer>
   *
   */
  controller?: MotionControllerProps;
}

export interface MotionChainProps extends GeneralHTMLAttributes {
  /**
   * @description
   * Defines properties that can be mandatoryly used across MP components.
   * It includes animation modes, transitions, delays, and durations. Basically
   * everything you need to make the web better :)
   *
   * @type {MotionAnimationProps[]} animations
   * @example
   *
   * const animations = Array.from({ length: 5 }, () => ({
   *    mode: ["scaleZoomIn","fadeIn"],
   *    transition: "slowSmooth",
   * }))
   * <MotionChain
   *   elementType="div"
   *   animations={animations}
   *   className="your-css-goes-here"
   * />
   *
   */
  animations: MotionAnimationProps[];
  /**
   * @description
   * The elementType prop allows you to specify the HTML element type
   * that will be used as the root element for the MotionContainer where
   * MotionChain encapsulates as its children.
   *
   * @default "div"
   * @type {React.ElementType}
   */
  elementType: React.ElementType;
  /**
   * @description
   * The children prop is a React nodes that will be rendered inside
   * the MotionChain component.
   *
   * @default undefined
   * @type {React.ReactNode[]}
   */
  children: React.ReactNode[];
  config: MotionChainConfigProps;
  /**
   * @description
   * This is the central scaffold part that you might see as
   * a god of CAS(Centralized Animation System). It's actually
   * standing on the top of each MP components to control
   * and manage the animation process by managing the flow.
   *
   * Highly recommended to use with both:
   * @type {UseAnimationControlProps}
   * @type {UseAnimationProps}
   *
   * @default undefined
   * @typedef {Object} MotionControllerProps
   * @param {boolean} isAnimationStopped
   * @param {boolean} reverseAnimation
   * @param {Partial<UseInViewOptions>} [configView]
   * @param {boolean} [trigger]
   *
   * @example
   *
   * //full example
   *
   * const { onReverse, control } = useAnimationControl();
   * const { isAnimationStopped, reverse } = useAnimation(control);
   *
   * // trigger the animation freely
   * <button onClick={onReverse}>Reverse</button>
   * <MotionContainer
   *   elementType="div"
   *   animation={{
   *     mode: ["fadeUp", "translate3dIn"],
   *     transition: "cubicBounce",
   *     duration: 1,
   *   }}
   *   controller={{
   *     trigger: true, // immediately start animation bypassing configView
   *     configView: { once: true, amount: 0.5 } // determine whether to start animation when it's in viewport
   *     isAnimationStopped, // pass the CAS props
   *     reverse // pass the CAS props
   *   }}
   *   className="your-css-goes-here"
   * />
   *   <IfThereIsChildComponent />
   * </MotionContainer>
   *
   */
  controller?: MotionControllerProps;
}
// ...
export interface MotionImageProps {
  /**
   * Defines properties that can be
   * mandatoryly used across MP components. It includes animation modes,
   * transitions, delays, and durations. Basically everything you need to
   * make the web better :)
   *
   * @typedef {Object} MotionAnimationProps
   * @property {AnimationKeys | AnimationKeys[]} mode - animation mode(s)
   * @property {TransitionKeys} transition - animation transition type
   * @property {number | undefined} [delay] - animation delay
   * @property {number} [duration] - animation duration
   *
   * @example
   *
   * <MotionContainer
   *   elementType="div"
   *   animation={{
   *     mode: ["fadeIn", "filterBlurIn"],
   *     transition: "smooth",
   *   }}
   *   className="your-css-goes-here"
   * />
   *
   */
  animation: MotionAnimationProps;
  /**
   * @description
   * A fallback component to be rendered when the image
   * is being loaded.
   *
   * @default undefined
   * @type {React.ReactNode}
   */
  fallback?: React.ReactNode;
  /**
   * @description
   * The wrapperClassName prop is a string that
   * specifies the class name(s) to be applied to the
   * wrapper element that wraps the child text elements
   *
   * @default undefined
   * @type {string}
   */
  wrapperClassName?: string;
  config: MotionImageConfigProps;
  /**
   * @description
   * This is the central scaffold part that you might see as
   * a god of CAS(Centralized Animation System). It's actually
   * standing on the top of each MP components to control
   * and manage the animation process by managing the flow.
   *
   * Highly recommended to use with both
   * @type {UseAnimationControlProps}
   * and @type {UseAnimationProps}
   *
   * @default undefined
   * @typedef {Object} MotionControllerProps
   * @param {boolean} isAnimationStopped
   * @param {boolean} reverseAnimation
   * @param {Partial<UseInViewOptions>} [configView]
   * @param {boolean} [trigger]
   *
   * @example
   *
   * //full example
   *
   * const { onReverse, control } = useAnimationControl();
   * const { isAnimationStopped, reverse } = useAnimation(control);
   *
   * // trigger the animation freely
   * <button onClick={onReverse}>Reverse</button>
   * <MotionContainer
   *   elementType="div"
   *   animation={{
   *     mode: ["fadeUp", "translate3dIn"],
   *     transition: "cubicBounce",
   *     duration: 1,
   *   }}
   *   controller={{
   *     trigger: true, // immediately start animation bypassing configView
   *     configView: { once: true, amount: 0.5 } // determine whether to start animation when it's in viewport
   *     isAnimationStopped, // pass the CAS props
   *     reverse // pass the CAS props
   *   }}
   *   className="your-css-goes-here"
   * />
   *   <IfThereIsChildComponent />
   * </MotionContainer>
   *
   */
  controller?: MotionControllerProps;
  imageLoading?: "eager" | "lazy";
  className?: string;
}

export interface MotionMovieProps extends GeneralHTMLAttributes {
  /**
   * @description
   * A superset of @type {MotionImageProps} that enables you
   * to create a visual slider conventionally using almost the
   * same syntax similar to @type {MotionImageProps}.
   *
   *
   * @default undefined
   * @type {MotionMovieAnimationsProps}
   */
  animations: MotionMovieAnimationsProps;
  /**
   * @description
   * This is the central scaffold part that you might see as
   * a god of CAS(Centralized Animation System). It's actually
   * standing on the top of each MP components to control
   * and manage the animation process by managing the flow.
   *
   * Highly recommended to use with both
   * @type {UseAnimationControlProps}
   * and @type {UseAnimationProps}
   *
   * @default undefined
   * @typedef {Object} MotionControllerProps
   * @param {boolean} isAnimationStopped
   * @param {boolean} reverseAnimation
   * @param {Partial<UseInViewOptions>} [configView]
   * @param {boolean} [trigger]
   *
   * @example
   *
   * //full example
   *
   * const { onReverse, control } = useAnimationControl();
   * const { isAnimationStopped, reverse } = useAnimation(control);
   *
   * // trigger the animation freely
   * <button onClick={onReverse}>Reverse</button>
   * <MotionContainer
   *   elementType="div"
   *   animation={{
   *     mode: ["fadeUp", "translate3dIn"],
   *     transition: "cubicBounce",
   *     duration: 1,
   *   }}
   *   controller={{
   *     trigger: true, // immediately start animation bypassing configView
   *     configView: { once: true, amount: 0.5 } // determine whether to start animation when it's in viewport
   *     isAnimationStopped, // pass the CAS props
   *     reverse // pass the CAS props
   *   }}
   *   className="your-css-goes-here"
   * />
   *   <IfThereIsChildComponent />
   * </MotionContainer>
   *
   */
  controller?: MotionControllerProps;
  config: MotionMovieConfigProps;
  /**
   * @description
   * A fallback component to be rendered when the image
   * is being loaded.
   *
   * @default undefined
   * @type {React.ReactNode}
   */
  fallback?: React.ReactNode;
  /**
   * @description
   * The wrapperClassName prop is a string that
   * specifies the class name(s) to be applied to the
   * wrapper element that wraps the child text elements
   *
   * @default undefined
   * @type {string}
   */
  wrapperClassName?: string;
}

export interface MotionLinkProps {
  timer: number;
  href: string;
  className?: string;
  onReverse: () => void;
  children: React.ReactNode;
}